# restful/docs/swagger.yaml
definitions:
...
  models.SystemUserTask:
    properties:
      addTime:
        description: 新增时间
        type: integer
      id:
        type: integer
      illustrate:
        description: 任务说明
        type: string
      isMust:
        description: 是否务必达成任务,1务必达成,0=满足其一
        type: integer
      isShow:
        description: 是否显示
        type: integer
      levelId:
        description: 等级id
        type: integer
      name:
        description: 任务名称
        type: string
      number:
        description: 限定数
        type: integer
      realName:
        description: 配置原名
        type: string
      sort:
        description: 排序
        type: integer
      taskType:
        description: 任务类型
        type: string
    type: object
...
  models.UserAddress:
    properties:
      addTime:
        description: 添加时间
        type: integer
      city:
        description: 收货人所在市
        type: string
      detail:
        description: 收货人详细地址
        type: string
      district:
        description: 收货人所在区
        type: string
      id:
        description: 用户地址id
        type: integer
      isDefault:
        description: 是否默认
        type: integer
      isDel:
        description: 是否删除
        type: integer
      latitude:
        description: 纬度
        type: string
      longitude:
        description: 经度
        type: string
      phone:
        description: 收货人电话
        type: string
      postCode:
        description: 邮编
        type: integer
      province:
        description: 收货人所在省
        type: string
      realName:
        description: 收货人姓名
        type: string
      uid:
        description: 用户id
        type: integer
    type: object
...
  models.UserLevel:
    properties:
      addTime:
        description: 添加时间
        type: integer
      discount:
        description: 享受折扣
        type: integer
      grade:
        description: 会员等级
        type: integer
      id:
        type: integer
      isDel:
        description: 是否删除,0=未删除,1=删除
        type: integer
      isForever:
        description: 是否永久
        type: integer
      levelId:
        description: 等级vip
        type: integer
      mark:
        description: 备注
        type: string
      merId:
        description: 商户id
        type: integer
      remind:
        description: 是否已通知
        type: integer
      status:
        description: 0:禁止,1:正常
        type: integer
      uid:
        description: 用户uid
        type: integer
      validTime:
        description: 过期时间
        type: integer
    type: object
...
info:
  contact: {}
  license: {}
paths:
...
  /systemUserTasks:
    get:
      parameters:
      - description: query, ?query=age:>:21,name:like:%jason%
        in: query
        name: query
        type: string
      - description: order, ?order=age:desc,created_at:asc
        in: query
        name: order
        type: string
      - description: page
        in: query
        name: page
        type: integer
      - description: pageSize
        in: query
        name: pageSize
        type: integer
      responses:
        "200":
          description: systemUserTask array
          schema:
            items:
              $ref: '#/definitions/models.SystemUserTask'
            type: array
      summary: List
      tags:
      - SystemUserTask
    post:
      parameters:
      - description: SystemUserTask
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.SystemUserTask'
      responses:
        "200":
          description: OK
          schema:
            type: string
      summary: Create
      tags:
      - SystemUserTask
  /systemUserTasks/{systemUserTaskId}:
    delete:
      parameters:
      - description: systemUserTaskId
        in: path
        name: systemUserTaskId
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            type: string
      summary: Delete
      tags:
      - SystemUserTask
    get:
      parameters:
      - description: systemUserTaskId
        in: path
        name: systemUserTaskId
        required: true
        type: string
      responses:
        "200":
          description: systemUserTask object
          schema:
            $ref: '#/definitions/models.SystemUserTask'
      summary: Get
      tags:
      - SystemUserTask
    patch:
      parameters:
      - description: systemUserTask
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.SystemUserTask'
      - description: systemUserTaskId
        in: path
        name: systemUserTaskId
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            type: string
      summary: Patch
      tags:
      - SystemUserTask
    put:
      parameters:
      - description: systemUserTask
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.SystemUserTask'
      - description: systemUserTaskId
        in: path
        name: systemUserTaskId
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            type: string
      summary: Put
      tags:
      - SystemUserTask
...
  /userAddresss:
    get:
      parameters:
      - description: query, ?query=age:>:21,name:like:%jason%
        in: query
        name: query
        type: string
      - description: order, ?order=age:desc,created_at:asc
        in: query
        name: order
        type: string
      - description: page
        in: query
        name: page
        type: integer
      - description: pageSize
        in: query
        name: pageSize
        type: integer
      responses:
        "200":
          description: userAddress array
          schema:
            items:
              $ref: '#/definitions/models.UserAddress'
            type: array
      summary: List
      tags:
      - UserAddress
    post:
      parameters:
      - description: UserAddress
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.UserAddress'
      responses:
        "200":
          description: OK
          schema:
            type: string
      summary: Create
      tags:
      - UserAddress
  /userAddresss/{userAddressId}:
    delete:
      parameters:
      - description: userAddressId
        in: path
        name: userAddressId
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            type: string
      summary: Delete
      tags:
      - UserAddress
    get:
      parameters:
      - description: userAddressId
        in: path
        name: userAddressId
        required: true
        type: string
      responses:
        "200":
          description: userAddress object
          schema:
            $ref: '#/definitions/models.UserAddress'
      summary: Get
      tags:
      - UserAddress
    patch:
      parameters:
      - description: userAddress
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.UserAddress'
      - description: userAddressId
        in: path
        name: userAddressId
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            type: string
      summary: Patch
      tags:
      - UserAddress
    put:
      parameters:
      - description: userAddress
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.UserAddress'
      - description: userAddressId
        in: path
        name: userAddressId
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            type: string
      summary: Put
      tags:
      - UserAddress
...
  /userLevels:
    get:
      parameters:
      - description: query, ?query=age:>:21,name:like:%jason%
        in: query
        name: query
        type: string
      - description: order, ?order=age:desc,created_at:asc
        in: query
        name: order
        type: string
      - description: page
        in: query
        name: page
        type: integer
      - description: pageSize
        in: query
        name: pageSize
        type: integer
      responses:
        "200":
          description: userLevel array
          schema:
            items:
              $ref: '#/definitions/models.UserLevel'
            type: array
      summary: List
      tags:
      - UserLevel
    post:
      parameters:
      - description: UserLevel
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.UserLevel'
      responses:
        "200":
          description: OK
          schema:
            type: string
      summary: Create
      tags:
      - UserLevel
  /userLevels/{userLevelId}:
    delete:
      parameters:
      - description: userLevelId
        in: path
        name: userLevelId
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            type: string
      summary: Delete
      tags:
      - UserLevel
    get:
      parameters:
      - description: userLevelId
        in: path
        name: userLevelId
        required: true
        type: string
      responses:
        "200":
          description: userLevel object
          schema:
            $ref: '#/definitions/models.UserLevel'
      summary: Get
      tags:
      - UserLevel
    patch:
      parameters:
      - description: userLevel
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.UserLevel'
      - description: userLevelId
        in: path
        name: userLevelId
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            type: string
      summary: Patch
      tags:
      - UserLevel
    put:
      parameters:
      - description: userLevel
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.UserLevel'
      - description: userLevelId
        in: path
        name: userLevelId
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            type: string
      summary: Put
      tags:
      - UserLevel
...
swagger: "2.0"